feat: API 7번 구현

+++ src/main/java/io/assemblers/project100endgame/profile/dto/ProfileResponse.java
@@ -4,7 +4,7 @@ import io.assemblers.project100endgame.profile.entity.UserProfile;
 
 public record ProfileResponse(
 	Long profileId,
-	Long level,
+	Integer level,
 	Long exp,
 	Long totalPlaySeconds
 ) {
